requirements.txt
acme==0.31.0
adal==1.2.7
aioredis==1.3.1
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.4.1
asn1crypto==0.22.0
async-timeout==4.0.2
attrs==19.3.0
autobahn==21.2.1
Automat==20.2.0
autopep8==1.6.0
azure==4.0.0
azure-applicationinsights==0.1.1
azure-batch==4.1.3
azure-common==1.1.28
azure-core==1.22.1
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-datalake-store==0.0.52
azure-eventgrid==1.3.0
azure-graphrbac==0.40.0
azure-keyvault==1.1.0
azure-loganalytics==0.1.1
azure-mgmt==4.0.0
azure-mgmt-advisor==1.0.1
azure-mgmt-applicationinsights==0.1.1
azure-mgmt-authorization==0.50.0
azure-mgmt-batch==5.0.1
azure-mgmt-batchai==2.0.0
azure-mgmt-billing==0.2.0
azure-mgmt-cdn==3.1.0
azure-mgmt-cognitiveservices==3.0.0
azure-mgmt-commerce==1.0.1
azure-mgmt-compute==4.6.2
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==1.5.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-containerservice==4.4.0
azure-mgmt-cosmosdb==0.4.1
azure-mgmt-datafactory==0.6.0
azure-mgmt-datalake-analytics==0.6.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==1.0.0
azure-mgmt-devspaces==0.1.0
azure-mgmt-devtestlabs==2.2.0
azure-mgmt-dns==2.1.0
azure-mgmt-eventgrid==1.0.0
azure-mgmt-eventhub==2.6.0
azure-mgmt-hanaonazure==0.1.1
azure-mgmt-iotcentral==0.1.0
azure-mgmt-iothub==0.5.0
azure-mgmt-iothubprovisioningservices==0.2.0
azure-mgmt-keyvault==1.1.0
azure-mgmt-loganalytics==0.2.0
azure-mgmt-logic==3.0.0
azure-mgmt-machinelearningcompute==0.4.1
azure-mgmt-managementgroups==0.1.0
azure-mgmt-managementpartner==0.1.1
azure-mgmt-maps==0.1.0
azure-mgmt-marketplaceordering==0.1.0
azure-mgmt-media==1.0.1
azure-mgmt-monitor==0.5.2
azure-mgmt-msi==0.2.0
azure-mgmt-network==2.7.0
azure-mgmt-notificationhubs==2.1.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==0.1.0
azure-mgmt-powerbiembedded==2.0.0
azure-mgmt-rdbms==1.9.0
azure-mgmt-recoveryservices==0.3.0
azure-mgmt-recoveryservicesbackup==0.3.0
azure-mgmt-redis==5.0.0
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.2.1
azure-mgmt-resource==2.2.0
azure-mgmt-scheduler==2.0.0
azure-mgmt-search==2.1.0
azure-mgmt-servicebus==0.5.3
azure-mgmt-servicefabric==0.2.0
azure-mgmt-signalr==0.1.1
azure-mgmt-sql==0.9.1
azure-mgmt-storage==2.0.0
azure-mgmt-subscription==0.2.0
azure-mgmt-trafficmanager==0.50.0
azure-mgmt-web==0.35.0
azure-nspkg==3.0.2
azure-servicebus==0.21.1
azure-servicefabric==6.3.0.0
azure-servicemanagement-legacy==0.20.7
azure-storage-blob
azure-storage-common==1.4.0
azure-storage-file==1.4.0
azure-storage-nspkg==3.1.0
azure-storage-queue==1.4.0
backcall==0.2.0
bleach==3.1.5
blinker==1.3
CacheControl==0.12.10
cachetools==4.2.4
certbot==0.31.0
certbot-nginx==0.31.0
certifi==2020.6.20
cffi==1.14.1
channels==3.0.4
channels-redis==3.3.1
chardet==2.3.0
charset-normalizer==2.0.12
click==8.0.4
ConfigArgParse==0.11.0
configobj==5.0.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
daphne==3.0.2
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.13
diff-match-patch==20200713
distlib==0.3.1
Django==2.2.24
django-cors-headers==3.10.1
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.4
django-extensions==3.1.5
django-filter==21.1
django-import-export==2.7.1
django-ipware==4.0.2
django-leaflet==0.19.0
django-phone-field==1.8.1
django-phonenumber-field==6.0.0
django-prometheus==2.2.0
django-rest-swagger==2.2.0
django-silk==4.2.0
django-simple-search==1.0.2
django-storages==1.12.3
django-vote==2.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
dnspython==2.2.0
drf-api-tracking==1.8.0
drf-extra-fields==3.2.1
drf-tracking==1.5.0
email-validator==1.1.3
entrypoints==0.3
et-xmlfile==1.1.0
factory-boy==3.0.1
Faker==4.1.2
filelock==3.0.12
firebase-admin==5.2.0
funcsigs==0.4
future==0.15.2
geocoder==1.38.1
google-api-core==2.5.0
google-api-python-client==2.38.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.0.0
google-crc32c==1.3.0
google-resumable-media==2.3.0
googleapis-common-protos==1.55.0
gprof2dot==2021.2.21
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.1.0
haversine==2.6.0
hiredis==2.0.0
httplib2==0.20.4
hyperlink==21.0.0
idna==2.5
importlib-metadata==4.8.3
importlib-resources==3.0.0
incremental==21.3.0
ipykernel==5.3.4
ipython==7.9.0
ipython-genutils==0.2.0
isodate==0.6.1
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
josepy==1.1.0
json5==0.9.5
jsonpatch==1.10
jsonpointer==1.9
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
jupyterlab==2.2.4
jupyterlab-server==1.2.0
Markdown==3.3.6
MarkupPy==1.14
MarkupSafe==0.23
mistune==0.8.4
mock==1.3.0
msgpack==1.0.3
msrest==0.6.21
msrestazure==0.6.4
nbconvert==5.6.1
nbformat==5.0.7
ndg-httpsclient==0.4.2
notebook==6.1.1
numpy==1.18.5
oauthlib==3.2.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.9
packaging==20.4
pandocfilters==1.4.2
parsedatetime==2.4
parso==0.7.1
pbr==1.8.0
pexpect==4.8.0
phonenumbers==8.12.6
pickleshare==0.7.5
Pillow==8.4.0
pipenv==2020.8.13
#pkg_resources==0.0.0
prettytable==0.7.2
prometheus-client==0.8.0
prompt-toolkit==2.0.10
proto-plus==1.20.3
protobuf==3.19.4
psycopg2==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.20
pydotplus==2.0.2
Pygments==2.6.1
pygraphviz==1.6
PyICU==2.8.1
PyJWT==1.3.0
pyOpenSSL==17.3.0
pyparsing==2.4.7
pyRFC3339==1.0
pyrsistent==0.16.0
pyserial==3.0.1
pyshp==2.1.0
python-dateutil==2.8.1
python-http-client==3.3.6
pytz==2014.10
PyYAML==3.11
pyzmq==19.0.2
ratelim==0.1.6
redis==4.3.4
redis-server==6.0.9
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.8
Send2Trash==1.5.0
sendgrid==6.9.6
sentry-sdk==1.5.6
serpy==0.3.1
service-identity==21.1.0
Shapely==1.7.0
simplejson==3.17.6
six==1.15.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
tablib==3.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
toml==0.10.2
tornado==6.0.4
traitlets==4.3.3
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.0.26
virtualenv-clone==0.5.4
walrus==0.9.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
zipp==1.2.0
zope.component==4.3.0
zope.event==4.2.0
zope.interface==5.4.0
